+++ packages/app/src/components/time-series-chart/components/timeline/index.ts
@@ -1,3 +1,4 @@
 export * from './timeline';
 export * from './components/marker';
 export * from './components/event-highlight';
+export type { TimelineEventConfig } from './logic';
