+++ packages/app/src/components/time-series-chart/components/tooltip/types.ts
@@ -2,9 +2,9 @@ import { TimestampedValue } from '@corona-dashboard/common';
 import {
   DataOptions,
   SeriesConfig,
-  TimelineEventConfig,
   TimespanAnnotationConfig,
 } from '../../logic';
+import { TimelineEventConfig } from '../timeline';
 
 export type TooltipData<T extends TimestampedValue> = {
   value: T;
@@ -28,15 +28,11 @@ export type TooltipData<T extends TimestampedValue> = {
   options: DataOptions;
 
   /**
-   * When hovering a date span annotation, the tooltip needs to know about it so
-   * that it can render the label accordingly. I am assuming here that we won't
-   * ever define overlapping annotations for now.
+   * When hovering a date-span/timeline annotation, the tooltip needs to know
+   * about it so that it can render the label accordingly. I am assuming here
+   * that we won't ever define overlapping annotations for now.
    */
   timespanAnnotation?: TimespanAnnotationConfig;
-
-  /**
-   * @TODO documentation
-   */
   timelineEvent?: TimelineEventConfig;
 
   /**
